Inline module instance behaviour, remove behaviours module
