--- lib/toy/behaviours.rb
@@ -1,48 +1,0 @@
-module Toy
-  module Behaviours
-    module Constants
-      def const_get(name)
-        name = name.to_sym
-        ::Kernel.raise ::NameError, "uninitialized constant #{inspect}::#{name}" unless constant_map.key?(name)
-
-        constant_map[name]
-      end
-
-      def const_set(name, value)
-        name = name.to_sym
-        ::Kernel.raise ::NameError unless name.match?(/^[A-Z][a-zA-Z_]*$/)
-
-        ::Kernel.warn "already initialized constant #{inspect}::#{name}" if constant_map.key?(name)
-
-        constant_map[name] = value
-      end
-
-      def constants
-        constant_map.keys.map(&:to_sym)
-      end
-
-      private
-
-      def constant_map
-        @constant_map ||= {}
-      end
-    end
-
-    module Methods
-      def define_method(name, method)
-        name = name.to_sym
-        method_map[name] = method
-      end
-
-      def instance_methods
-        method_map.keys.map(&:to_sym)
-      end
-
-      private
-
-      def method_map
-        @method_map ||= {}
-      end
-    end
-  end
-end
